# Improve fixture customization performance by adding laziness to adapted collection