@zvirja 2017-09-07
@adamchester Given that this PR is all about performance, do you know whether you will be able to allocate some time in the nearby future and review it? As you can see, no tests failed, so logically nothing changed.

I'll gain a benefit from it in my on-day basis activity so I'm a bit impatient to release this stuff 😉