@zvirja 2017-09-08
@moodmosaic  Adam seems to be absent - could you please take a look (unless you explicitly decided to skip this one) if possible?