@moodmosaic 2017-09-18
@zvirja, I apologize for the being late here.

Perhaps, in addition to what I wrote [here](https://github.com/AutoFixture/AutoFixture/issues/703#issuecomment-330116357), you could also go ahead and merge these perf-related changes on your own―as long as no breaking changes are introduced.

👍 :1st_place_medal: 💯 🐴 🚀